Rewrote autoencoder model for MB_particle dataset, modified forward function to perform umbrella sampling
